refactor(base): Moved static methods to Requests
