+++ base/src/main/java/io/fabric8/launcher/base/http/HttpClient.java
@@ -16,7 +16,6 @@ import javax.net.ssl.X509TrustManager;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import io.fabric8.launcher.base.JsonUtils;
-import io.fabric8.launcher.base.identity.Identity;
 import okhttp3.Call;
 import okhttp3.Callback;
 import okhttp3.Dispatcher;
@@ -172,27 +171,6 @@ public class HttpClient {
                 }
             }
     };
-
-    /**
-     * Request builder factory prepared with identity authorization
-     *
-     * @param identity          the {@link Identity} to use
-     * @param authorizationType the {@link AuthorizationType} to use
-     * @return the {@link Request.Builder}
-     */
-    public static Request.Builder securedRequest(final Identity identity, final AuthorizationType authorizationType) {
-        return new Request.Builder().header(authorizationType.getHeaderName(), identity.toRequestAuthorization(authorizationType));
-    }
-
-    /**
-     * Request builder factory prepared with default identity authorization {@link Identity#toRequestAuthorization()}
-     *
-     * @param identity the {@link Identity} to use
-     * @return the {@link Request.Builder}
-     */
-    public static Request.Builder securedRequest(final Identity identity) {
-        return new Request.Builder().header(identity.getDefaultAuthorizationType().getHeaderName(), identity.toRequestAuthorization());
-    }
 
     private static OkHttpClient createClient(@Nullable ExecutorService executorService) {
         OkHttpClient.Builder builder = new OkHttpClient.Builder();
