+++ services/keycloak-service-impl/src/main/java/io/fabric8/launcher/service/keycloak/impl/KeycloakServiceImpl.java
@@ -19,7 +19,7 @@ import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.Response;
 
-import static io.fabric8.launcher.base.http.HttpClient.securedRequest;
+import static io.fabric8.launcher.base.http.Requests.securedRequest;
 import static java.util.Objects.requireNonNull;
 
 /**
